requirements.txt
psutil
absl-py==0.7.1
astor==0.8.0
boto==2.49.0
boto3==1.9.245
botocore==1.12.245
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
deepdiff==4.0.7
Django==2.2.4
django-site-maintenance==0.1.3
django-stubs==1.2.0
djangorestframework==3.10.2
djangorestframework-stubs==1.0.0
docutils==0.15.2
gast==0.2.2
gensim==3.8.1
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
imgaug==0.2.9
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.13.2
jsonpickle==1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mock==3.0.5
mpld3==0.3
mypy==0.720
mypy-extensions==0.4.2
networkx==2.3
numexpr==2.7.0
numpy==1.16.0
opencv-python==4.1.0.25
ordered-set==3.1.1
pandas==0.25.0
Pillow==6.1.0
plotly==4.1.1
protobuf==3.9.0
psycopg2==2.8.3
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
requests==2.22.0
retrying==1.3.3
ruamel.appconfig==0.5.4
ruamel.std.argparse==0.8.1
s3transfer==0.2.1
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.0
seaborn==0.9.0
Shapely==1.6.4.post2
six==1.12.0
sklearn==0.0
smart-open==1.8.4
SQLAlchemy==1.3.6
sqlparse==0.3.0
tables==3.5.2
tb-nightly==1.14.0a20190603
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.6
watchtower==0.7.3
Werkzeug==0.15.5
wordcloud==1.5.0
wrapt==1.11.2
